[COOK-2873] Allow "blank" values in generated rsyncd.conf
When writing rsyncd.conf, the previous code skipped entries that
satisfied value.blank?, in order to avoid writing out unset attributes.
But this also affected values set to false where the built-in default of
rsync is true.  For example, it was impossible to configure an rsync
server with "read only = false".  Fix this by testing for nil? instead.